ADE7880  controlled by Raspberry Pi
This project was meant to control an electric motor attached to a diesel
engine to produce electricity. An ADE7880 interfaced with a Raspberry Pi
by means of SPI CS0 is used to control the energy production while a
Renesas M16C62P is used to control the Diesel motor by reading the
output signals of its sensors and serving the data to the Raspberry Pi
using SPI CS1. The gathered information is then made available to a
MicroScada like system by means of libIEC61850 library